Adding status for Halls and ads , adding new endpoints for the admin to approve/reject ads or halls

+++ ISSHAR.Application/Services/HallService.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using ISSHAR.Application.DTOs.HallDTOs;
 using ISSHAR.DAL.Entities;
+using ISSHAR.DAL.Enums;
 using ISSHAR.DAL.Repositories;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 
 namespace ISSHAR.Application.Services
@@ -9,23 +11,25 @@ namespace ISSHAR.Application.Services
     public class HallService : IHallService
     {
         private readonly IHallRepository _hallRepository;
+        private readonly IBookingRepository _bookingRepository;
         private readonly IMapper _mapper;
         private readonly ILogger<IHallService> _logger;
         private readonly IImageService _cloudinary;
 
-        public HallService(IHallRepository hallRepository, IMapper mapper, ILogger<IHallService> logger,IImageService cloudinary)
+        public HallService(IHallRepository hallRepository,IBookingRepository bookingRepository, IMapper mapper, ILogger<IHallService> logger,IImageService cloudinary)
         {
             _hallRepository = hallRepository;
+            _bookingRepository = bookingRepository;
             _mapper = mapper;
             _logger = logger;
             _cloudinary = cloudinary;
         }
 
-        public async Task<ICollection<HallDisplayDTO>> GetAllAsync()
+        public async Task<ICollection<HallDisplayDTO>> GetHallsByStatusAsync(Status status)
         {
             try
             {
-                var halls = await _hallRepository.GetAllAsync();
+                var halls = await _hallRepository.GetByStatusAsync(status);
                 var hallDtos = _mapper.Map<ICollection<HallDisplayDTO>>(halls);
                 return hallDtos;
             }
@@ -51,19 +55,16 @@ namespace ISSHAR.Application.Services
             }
         }
 
-        public async Task <HallDisplayDTO> AddAsync(HallDTO hallDTO)
+        public async Task<HallDisplayDTO> AddAsync(HallDTO hallDTO)
         {
             try
             {
                 var hall = _mapper.Map<Hall>(hallDTO);
-                var hallImages = new List<HallImage>();
-                foreach (var file in hallDTO.ImagesAsFiles)
-                {
-                    hallImages.Add(new HallImage { ImageUrl = await _cloudinary.UploadImageAsync(file)});
-                }
-                hall.HallImages= hallImages;
-                hall.Logo = await _cloudinary.UploadImageAsync(hallDTO.LogoFile);
-                await _hallRepository.AddAsync(hall);                
+                hall.HallImages = await UploadHallImagesAsync(hallDTO.ImagesAsFiles.ToList());
+                hall.Logo = await GetHallLogoUrl(hallDTO.LogoFile);
+                hall.Status = Status.Pending;
+                await _hallRepository.AddAsync(hall);
+
                 var hallDisplayDTO = _mapper.Map<HallDisplayDTO>(hall);
                 return hallDisplayDTO;
             }
@@ -103,6 +104,16 @@ namespace ISSHAR.Application.Services
                 {
                     return false;
                 }
+                var hasFutureBookings = await _bookingRepository.HasFutureBookingsAsync(id);
+                if(hasFutureBookings)
+                {
+                    return false;
+                }
+                var hallBookings = await _bookingRepository.GetByHallIdAsync(id);
+                foreach(var booking in hallBookings)
+                {
+                    await _bookingRepository.DeleteAsync(booking);
+                }
                 await _hallRepository.DeleteAsync(existingHall);
                 return true;
             }
@@ -127,6 +138,7 @@ namespace ISSHAR.Application.Services
                 throw;
             }
         }
+
         public async Task<ICollection<HallDisplayDTO>> GetFilteredHallsAsync(HallFitlerBody filter)
         {
             try
@@ -142,6 +154,44 @@ namespace ISSHAR.Application.Services
             }
         }
 
+        public async Task<bool> ChangeStatusAsync(int id, string newStatus)
+        {
+            try
+            {
+                if (!Enum.TryParse(newStatus, out Status status))
+                {
+                    return false;
+                }
+                var advertisement = await _hallRepository.GetByIdAsync(id);
+                if (advertisement == null)
+                {
+                    return false;
+                }
+                advertisement.Status = status;
+                await _hallRepository.UpdateAsync(advertisement);
 
+                return true;
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "Error occurred while changing status of advertisement with ID: {AdvertisementId}", id);
+                throw;
+            }
+        }
+
+        private async Task<List<HallImage>> UploadHallImagesAsync(List<IFormFile> imagesAsFiles)
+        {
+            var hallImages = new List<HallImage>();
+            foreach (var file in imagesAsFiles)
+            {
+                hallImages.Add(new HallImage { ImageUrl = await _cloudinary.UploadImageAsync(file) });
+            }
+            return hallImages;
+        }
+
+        private async Task<string> GetHallLogoUrl(IFormFile logoFile)
+        {
+            return logoFile == null ? DefaultImageUrls.HallLogoUrl : await _cloudinary.UploadImageAsync(logoFile);
+        }
     }
 }
